requirements.txt
numpy
pandas
scipy
python-dateutil==2.8.1
pytz==2021.1
hyperopt==0.2.5
certifi==2020.12.5
pyyaml==5.4.1
networkx==2.5.1
scikit-learn==0.24.2
keras
six==1.15.0
theano==1.0.3
psutil==5.8.0
pympler==0.9
tensorflow
tables
scikit-optimize==0.8.1
python-telegram-bot==13.5
tqdm
dill
pillow
ruamel.yaml
